Move API for coin to module HSChain.Examples.Coin.API
